pyproject.toml
[project]
name = "brick-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "bcrypt>=5.0.0",
    "boto3>=1.42.59",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "openai>=2.21.0",
    "pycups>=2.0.4",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sentry-sdk>=2.48.0",
    "uvicorn>=0.40.0",
]
